+++ ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/response/JsonRpcError.java
@@ -21,6 +21,7 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonGetter;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import org.apache.tuweni.bytes.Bytes;
 
 @JsonInclude(value = JsonInclude.Include.NON_NULL)
 @JsonFormat(shape = JsonFormat.Shape.OBJECT)
@@ -28,6 +29,7 @@ public class JsonRpcError {
   private final int code;
   private final String message;
   private final String data;
+  private String reason;
 
   @JsonCreator
   public JsonRpcError(
@@ -41,6 +43,15 @@ public class JsonRpcError {
 
   public JsonRpcError(final RpcErrorType errorType, final String data) {
     this(errorType.getCode(), errorType.getMessage(), data);
+
+    // For execution reverted errors decode the data (if present)
+    if (errorType == RpcErrorType.REVERT_ERROR && data != null) {
+      JsonRpcErrorResponse.decodeRevertReason(Bytes.fromHexString(data))
+          .ifPresent(
+              (decodedReason) -> {
+                this.reason = decodedReason;
+              });
+    }
   }
 
   public JsonRpcError(final RpcErrorType errorType) {
@@ -54,7 +65,7 @@ public class JsonRpcError {
 
   @JsonGetter("message")
   public String getMessage() {
-    return message;
+    return (reason == null ? message : message + ": " + reason);
   }
 
   @JsonGetter("data")
@@ -72,12 +83,12 @@ public class JsonRpcError {
     }
     final JsonRpcError that = (JsonRpcError) o;
     return code == that.code
-        && Objects.equals(message, that.message)
+        && Objects.equals(message.split(":", -1)[0], that.message.split(":", -1)[0])
         && Objects.equals(data, that.data);
   }
 
   @Override
   public int hashCode() {
-    return Objects.hash(code, message, data);
+    return Objects.hash(code, message.split(":", -1)[0], data);
   }
 }